requirements.txt
fastapi
uvicorn[standard]
python-pptx
python-dotenv
openai
qrcode
pillow
